requirements.txt
# requirements.txt
opensearch-py==2.4.2
sentence-transformers==2.7.0
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
markdown==3.5.1
pyyaml>=6.0.2
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
urllib3>=2.5.0,<3.0
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0